Move orchestration package, update README code
