+++ client/client.go
@@ -5,12 +5,12 @@ import (
 	"fmt"
 
 	"github.com/coinbase/staking-client-library-go/client/options"
-	orchestrationClient "github.com/coinbase/staking-client-library-go/client/orchestration/v1"
+	"github.com/coinbase/staking-client-library-go/client/orchestration"
 	"github.com/coinbase/staking-client-library-go/client/rewards"
 )
 
 type StakingClient struct {
-	Orchestration *orchestrationClient.OrchestrationServiceClient
+	Orchestration *orchestration.OrchestrationServiceClient
 	Rewards       *rewards.RewardsServiceClient
 }
 
@@ -19,7 +19,7 @@ func New(
 	ctx context.Context,
 	stakingOpts ...options.StakingClientOption,
 ) (*StakingClient, error) {
-	orchestrationClient, err := orchestrationClient.NewOrchestrationServiceClient(ctx, stakingOpts...)
+	orchestrationClient, err := orchestration.NewOrchestrationServiceClient(ctx, stakingOpts...)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create orchestration client: %w", err)
 	}
