+++ client/orchestration/network.go
@@ -1,4 +1,4 @@
-package v1
+package orchestration
 
 import (
 	"context"
