+++ client/orchestration/view_staking_context.go
@@ -1,4 +1,4 @@
-package v1
+package orchestration
 
 import (
 	"context"
