+++ examples/ethereum/create-workflow/main.go
@@ -1,12 +1,11 @@
 /*
- * This example code, demonstrates creating an Ethereum Partial ETH (<32 ETH) workflow, the fundamental process handler for an E2E staking experience.
+ * This example code demonstrates creating an Ethereum Partial ETH (<32 ETH) workflow, the fundamental process handler for an E2E staking experience.
  */
 
 package main
 
 import (
 	"context"
-	"fmt"
 	"log"
 
 	"github.com/coinbase/staking-client-library-go/auth"
@@ -16,28 +15,21 @@ import (
 )
 
 func main() {
-	// TODO: Add your project ID found at cloud.coinbase.com or in your API key.
-	projectID := ""
-
 	ctx := context.Background()
 
+	// Loads the API key from the default location.
 	apiKey, err := auth.NewAPIKey(auth.WithLoadAPIKeyFromFile(true))
 	if err != nil {
 		log.Fatalf("error loading API key: %s", err.Error())
 	}
 
-	// Create a staking client.
+	// Creates the Coinbase Staking API client.
 	stakingClient, err := client.New(ctx, options.WithAPIKey(apiKey))
 	if err != nil {
 		log.Fatalf("error instantiating staking client: %s", err.Error())
 	}
 
-	if projectID == "" {
-		log.Fatalf("projectID must be set")
-	}
-
 	req := &stakingpb.CreateWorkflowRequest{
-		Parent: fmt.Sprintf("projects/%s", projectID),
 		Workflow: &stakingpb.Workflow{
 			Action: "protocols/ethereum_kiln/networks/holesky/actions/stake",
 			StakingParameters: &stakingpb.Workflow_EthereumKilnStakingParameters{
